bumped djs to v13 added getgame.js and other small fixes
