+++ commands/addgames.js
@@ -1,5 +1,4 @@
-const fs = require('fs');
-const fetch = require('node-fetch');
+const getgame = require('../getgame');
 module.exports = {
 	name: 'addgames',
 	aliases: ['ags'],
@@ -22,30 +21,7 @@ module.exports = {
 		let games = [];
 		for (let x of argz) {
 			x = x.trim();
-			const errormsg = `are you sure https://www.speedrun.com/${x} exists\n||if it didn't work try deleting unnecessary spaces||`;
-			const res = await fetch(`https://www.speedrun.com/api/v1/games/${x}`);
-			let json = await res.json();
-			if (!json.data) {
-				const ares = await fetch(`https://www.speedrun.com/api/v1/games?name=${x}`);
-				let ajson = await ares.json();
-				if (!ajson.data) {
-					message.reply(errormsg);
-					continue;
-				}
-				if (ajson.data[0]) {
-					json.data = ajson.data.find(y => y.names.international == decodeURIComponent(x))
-					if (!json.data) {
-						message.reply(errormsg);
-						continue;
-					}
-				} else {
-					json = ajson;
-					if (!json.data.id) {
-						message.reply(errormsg);
-						continue;
-					}
-				}
-			}
+			let json = await getgame(x, message);
 			const [game,] = await Game.findOrCreate({
 				where: {
 				  id: json.data.id
