+++ package.json
@@ -10,7 +10,7 @@
   "author": "NoobJsPerson",
   "license": "ISC",
   "dependencies": {
-    "discord.js": "^12.5.1",
+    "discord.js": "^13.10.3",
     "dotenv": "10.0.0",
     "node-fetch": "^2.6.1",
     "sequelize": "^6.21.3",
